pyproject.toml
# SPDX-License-Identifier: MIT

[tool.poetry]
name = "neenee"
version = "0.0.1"
description = "An elegant & scalable Discord bot which aces in robustness (and in doing chores :D)"
authors = ["HitBlast <hitblastlive@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/hitblast/Neenee"
packages = [
    { include = "neenee" }
]

[tool.poetry.dependencies]
python = "^3.12"
disnake = "^2.9.1"
rich = "^13.7.1"
click = "^8.1.7"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"

[tool.poetry.scripts]
neenee = "neenee.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["venv"]
